add configs directory and rename celeryconfig to celery_config
