+++ setup.cfg
@@ -6,8 +6,7 @@ known_first_party =
     ml_model
     tasks
     utils
-    config
-    celeryconfig
+    configs
     enums
     payloads
