+++ worker/worker.py
@@ -1,8 +1,8 @@
 from celery import Celery
 from redis import Redis
 
-import celeryconfig
-from config import celery_worker_settings, redis_settings
+import configs.celery_config
+from configs.config import celery_worker_settings, redis_settings
 
 
 app = Celery(
@@ -11,7 +11,7 @@ app = Celery(
     backend=celery_worker_settings.backend_uri,
     include=["tasks"],
 )
-app.config_from_object(celeryconfig)
+app.config_from_object(configs.celery_config)
 redis = Redis(
     host=redis_settings.redis_host,
     port=redis_settings.redis_port,
